+++ lib/prototype_constrained_box.dart
@@ -1,1 +1,2 @@
+export 'src/rendering.dart';
 export 'src/widget.dart';